Feature/native tlm embed (#565)
Remove addons for TLM, they are in core lib now (baseline/<framework>/embeddings.py)
Post BERT small NPZ checkpoints from my dropbox location. These files were converted using the API example convert_huggingface_checkpoints.py and the SHA1 obtained with sha1sum.
In PyTorch TLM, which supports 3 types of checkpoints, test if PKZIP and treat as NPZ after testing if its a Transformers model
Add support to EmbeddingsDownloader not to unzip via a property unzip in the index (it defaults to true but should be set to false in the index for NPZ files)
Remove mead configs that are duplicate except backend
Closes #563
