--- contributors.md
@@ -1,5 +1,0 @@
-[<img alt="dpressel" src="https://avatars1.githubusercontent.com/u/247881?v=4&s=117" width="117">](https://github.com/dpressel)[<img alt="blester125" src="https://avatars3.githubusercontent.com/u/10950530?v=4&s=117" width="117">](https://github.com/blester125)[<img alt="sagnik" src="https://avatars3.githubusercontent.com/u/4563472?v=4&s=117" width="117">](https://github.com/sagnik)[<img alt="mtbarta" src="https://avatars0.githubusercontent.com/u/2423845?v=4&s=117" width="117">](https://github.com/mtbarta)[<img alt="amyhemmeter" src="https://avatars2.githubusercontent.com/u/23508540?v=4&s=117" width="117">](https://github.com/amyhemmeter)
-
-[<img alt="ijindal" src="https://avatars3.githubusercontent.com/u/19698647?v=4&s=117" width="117">](https://github.com/ijindal)
-
-
